Moved Subgraph and Node functions to new module
